Add reference links and documents
This reference links should provide a better understanding about this
game, what it is, what is expected and so on.

I have also found that there are plenty of references across the
internet, some of them no longer exist, so whenever I can find the
original source I will make a copy of it and add it to the
`docs/references` folder and hopefully still give credit to the original
author.