Add Mark Damon Hughes Game Design collection
When I added Mark's article 07 - that is directly related with the work
in this repository, I felt that the article would feel loose, as there
are 14 other articles around number 07.
For that reason, I've decided to add the remaining articles, with the
exception of number 14 which I felt it would be a bigger hassle to try
and find all the links to keep the article as original as possible.